==> 3. Analysis.Rmd <==
---
title: "3. Analysis"
output: html_document
date: "2024-04-03"
---

```{r setup, include = F}
####  Setup  ####
## Load packages
library(tidyverse)
library(broom)
library(here)
library(scales)
library(knitr)
library(kableExtra)
library(oddsratio)
`%+%` <- paste0


## Load data
df <- readRDS(here("Data", "Clean Data with Concept Breadth.rds"))



####  Analysis  ####
## Set output table, functions
out <- tibble()

my_lm <- function(iv, dv, rq, control = F) {
  
  formula_temp <- dv %+% " ~ " %+% iv
  
  if(control) formula_temp <- formula_temp %+% " + phq_sum"
  
  regression_temp <- lm(
    data = df,
    formula = formula_temp
  )
  
  output_temp <- tidy(regression_temp) %>%
    filter(! term %in% c("(Intercept)", "phq_sum")) %>%
    mutate(dv = dv, control = control, rq = rq) %>%
    select(rq, dv, iv = term, control, estimate, p.value)
  
  return(output_temp)
  
}

my_glm <- function(iv, dv, rq, control = F) {
  
  formula_temp <- dv %+% " ~ " %+% iv
  
  if(control) formula_temp <- formula_temp %+% " + phq_sum"
  
  regression_temp <- glm(
    data = df,
    formula = formula_temp,
    family = "binomial"
  )
  
  output_temp <- tidy(regression_temp) %>%
    filter(! term %in% c("(Intercept)", "phq_sum")) %>%
    mutate(estimate = exp(estimate), dv = dv, control = control, rq = rq) %>%
    select(rq, dv, iv = term, control, estimate, p.value)
  
  return(output_temp)
  
}


## Analysis 1A: What student characteristics are associated with depression self-labeling?
student_characteristics <- c(
  "age", "race_ethnicity", "gender", "sexuality", "cob_usa", "parent_cob_usa",
  "dep_dx", "dep_therapy", "dep_meds", "dep_fam", "dep_friends"
)

for(x in student_characteristics) {
  
  for(y in c("dep_self_id == 'Yes'", "dep_self_id %in% c('Yes', 'Not sure')")) {
    
    output_uncontrolled <- my_glm(iv = x, dv = y, rq = "1a", control = F)
    out <- bind_rows(out, output_uncontrolled)
  
    output_controlled <- my_glm(iv = x, dv = y, rq = "1a", control = T)
    out <- bind_rows(out, output_controlled)
    
  }

}


## Analysis 1B: Is depression self-labeling associated with certain coping strategies?
coping_strategies <- c("cbas_cn_sum", "cbas_cs_sum", "cbas_bn_sum", "cbas_bs_sum", "therapy_eff", "med_eff")

# Linear regressions for continuous outcomes
for(y in coping_strategies) {
  
  output_uncontrolled <- my_lm(iv = "dep_self_id", dv = y, rq = "1b", control = F)
  out <- bind_rows(out, output_uncontrolled)
  
  output_controlled <- my_lm(iv = "dep_self_id", dv = y, rq = "1b", control = T)
  out <- bind_rows(out, output_controlled)
  
}

# Logistic regressions for binary outcomes
output_uncontrolled <- my_glm(iv = "dep_self_id", dv = "ssi_opt_in", rq = "1b", control = F)
out <- bind_rows(out, output_uncontrolled)

output_controlled <- my_glm(iv = "dep_self_id", dv = "ssi_opt_in", rq = "1b", control = T)
out <- bind_rows(out, output_controlled)


## Analysis 2A: What student characteristics are associated with greater depression concept breadth?
for(x in student_characteristics) {
  
  output_uncontrolled <- my_lm(iv = x, dv = "concept_breadth", rq = "2a", control = F)
  out <- bind_rows(out, output_uncontrolled)
  
  output_controlled <- my_lm(iv = x, dv = "concept_breadth", rq = "2a", control = T)
  out <- bind_rows(out, output_controlled)
  
}


## Analysis 2B: Is greater depression concept breadth associated with depression self-labeling?
for(y in c("dep_self_id == 'Yes'", "dep_self_id %in% c('Yes', 'Not sure')")) {
  
  output_uncontrolled <- my_glm(iv = "concept_breadth", dv = y, rq = "2b", control = F)
  out <- bind_rows(out, output_uncontrolled)
  
  output_controlled <- my_glm(iv = "concept_breadth", dv = y, rq = "2b", control = T)
  out <- bind_rows(out, output_controlled)
  
}


## Adjust p-values for all tests
out$p.adj <- p.adjust(out$p.value, "BH")
```

## Sample Description

```{r}
# Summarize continuous variables
df %>%
  pivot_longer(c(age, phq_sum)) %>%
  group_by(name) %>%
  summarize(
    mean = mean(value),
    sd = sd(value),
    min = min(value),
    max = max(value)
  )

# Summarize categorical variables
count_with_percents <- function(x) {
  
  df %>%
    count({{x}}) %>%
    mutate(percent = percent(n / sum(n), .1)) %>%
    return()
  
}

count_with_percents(gender)
count_with_percents(race_ethnicity)
count_with_percents(sexuality)
count_with_percents(cob_usa)
count_with_percents(parent_cob_usa)
count_with_percents(dep_self_id)
count_with_percents(dep_dx)
count_with_percents(dep_therapy)
count_with_percents(dep_meds)
count_with_percents(dep_fam)
count_with_percents(dep_friends)
count_with_percents(phq_sum >= 3)
```

## Analysis 1: Depression Self-Labeling and Coping

### Analysis 1A: What student characteristics are associated with depression self-labeling?

```{r}
out %>%
  filter(
    rq == "1a", 
    dv == "dep_self_id == 'Yes'"
  ) %>%
  mutate(control = if_else(control, "controlled", "not_controlled")) %>%
  pivot_wider(
    id_cols = iv,
    names_from = control,
    values_from = c(estimate, p.adj)
  ) %>%
  select(iv, estimate_controlled, p.adj_controlled, estimate_not_controlled, p.adj_not_controlled) %>%
  kable() %>%
  kable_styling()
```

### Analysis 1B: Is depression self-labeling associated with certain coping strategies?

```{r}
out %>%
  filter(rq == "1b") %>%
  mutate(control = if_else(control, "controlled", "not_controlled")) %>%
  pivot_wider(
    id_cols = c(iv, dv),
    names_from = control,
    values_from = c(estimate, p.adj)
  ) %>%
  select(iv, dv, estimate_controlled, p.adj_controlled, estimate_not_controlled, p.adj_not_controlled) %>%
  kable() %>%
  kable_styling()
```

#### Figure 1

Effects of self-labeling on avoidant coping, controlling for depression symptom severity.

```{r}
figure_2_outcomes <- c("cbas_cn_sum", "cbas_cs_sum", "cbas_bn_sum", "cbas_bs_sum", "cbas_sum")

median_phq_sum <- median(df$phq_sum)

figure_2_data <- tibble()

for(y in figure_2_outcomes) {
  
  y_scaled <- scale(df[[y]])
  
  regression_temp <- lm(
    data = df,
    formula = as.formula(y %+% " ~ dep_self_id + phq_sum")
  )
  
  data_temp <- tibble(
    phq_sum = median_phq_sum,
    dep_self_id = c("Yes", "Not sure", "No")
  )
  
  predictions_temp <- predict(
    object = regression_temp,
    newdata = data_temp,
    type = "response",
    interval = "confidence"
  )
  
  out_temp <- bind_cols(
    tibble(
      y = y,
      dep_self_id = c("Yes", "Not sure", "No")
    ),
    predictions_temp
  )
  
  figure_2_data <- bind_rows(figure_2_data, out_temp)
  
}

figure_2_data %>%
  mutate(
    y = recode(
      y, 
      "cbas_cn_sum" = "Cognitive Non-Social Avoidance",
      "cbas_cs_sum" = "Cognitive Social Avoidance",
      "cbas_bn_sum" = "Behavioral Non-Social Avoidance",
      "cbas_bs_sum" = "Behavioral Social Avoidance",
      "cbas_sum" = "Total Avoidance"
    )
  ) %>%
  ggplot(aes(dep_self_id, fit)) +
  geom_point() +
  geom_errorbar(aes(ymin = lwr, ymax = upr)) +
  scale_x_discrete(name = "Depression Self-Labeling") +
  scale_y_continuous(name = "z-Score") +
  ggtitle("Avoidant Coping by Depression Self-Labeling", "Predicted Values Controlling for Depressive Symptom Severity") +
  theme_classic() +
  facet_wrap(~ y, scales = "free_y")
```

## Analysis 2: Depression Concept Breadth and Self-Labeling

### Analysis 2A: What student characteristics are associated with greater depression concept breadth?

```{r}
out %>%
  filter(
    rq == "2a"
  ) %>%
  mutate(control = if_else(control, "controlled", "not_controlled")) %>%
  pivot_wider(
    id_cols = iv,
    names_from = control,
    values_from = c(estimate, p.adj)
  ) %>%
  select(iv, estimate_controlled, p.adj_controlled, estimate_not_controlled, p.adj_not_controlled) %>%
  kable() %>%
  kable_styling()
```

### Analysis 2B: Is greater depression concept breadth associated with depression self-labeling?

Note: I did *not* pre-register that I would control for FDR here, since these are only four tests and they are at the core of the study (the "primary" RQs, I guess). That said, I probably should have. But there's some flexibility here.

```{r}
out %>%
  filter(rq == "2b") %>%
  mutate(control = if_else(control, "controlled", "not_controlled")) %>%
  pivot_wider(
    id_cols = dv,
    names_from = control,
    values_from = c(estimate, p.adj)
  ) %>%
  select(dv, estimate_controlled, p.adj_controlled, estimate_not_controlled, p.adj_not_controlled) %>%
  kable() %>%
  kable_styling()
```

#### Figure 2

Predicted self-labeling by concept breadth, controlling for depression symptom severity.

```{r}
self_label_y_by_concept_breadth_model <- glm(
  data = df,
  formula = dep_self_id == "Yes" ~ concept_breadth + phq_sum,
  family = "binomial"
)

self_label_y_by_concept_breadth_fitted_data <- expand.grid(
  phq_sum = 0:6,
  concept_breadth = seq(-3, 3, .05)
) %>%
  mutate(
    p_self_id = predict(
      self_label_y_by_concept_breadth_model,
      newdata = .,
      type = "response"
    ),
    self_id = "Yes"
  )

self_label_ym_by_concept_breadth_model <- glm(
  data = df,
  formula = dep_self_id %in% c("Yes", "Not sure") ~ concept_breadth + phq_sum,
  family = "binomial"
)

self_label_ym_by_concept_breadth_fitted_data <- expand.grid(
  phq_sum = 0:6,
  concept_breadth = seq(-3, 3, .05)
) %>%
  mutate(
    p_self_id = predict(
      self_label_ym_by_concept_breadth_model,
      newdata = .,
      type = "response"
    ),
    self_id = "Yes/Not sure"
  )

fitted_data_combined <- bind_rows(self_label_y_by_concept_breadth_fitted_data, self_label_ym_by_concept_breadth_fitted_data)

fitted_data_combined %>%
  filter(self_id == "Yes") %>% # Can also try faceted graph, but it looks confusing - even vertically
  ggplot() +
  geom_line(aes(concept_breadth, p_self_id, color = as.character(phq_sum)),
            linewidth = 1) +
  scale_x_continuous(
    name = "Depression Concept Breadth (z-Score)"
  ) +
  scale_y_continuous(
    name = "Probability of Depression Self-Labeling",
    labels = percent_format(),
    limits = c(0, 1)
  ) +
  scale_color_discrete(
    name = "PHQ-2 Sum Score",
    guide = guide_legend(reverse = T)
  ) +
  # facet_wrap(~ self_id, ncol = 1) +
  theme_classic()
```

## Etc.

### Quick test - do LGB+ students *over*identify? (No! They just *under*identify *less*)

```{r}
df %>%
  group_by(sexuality) %>%
  summarize(
    self_label = mean(dep_self_id == "Yes"),
    phq_screen = mean(phq_sum >= 3)
  )

# FYI - they also have more depression literacy
summary(lm(df$dlit_sum ~ df$sexuality))
```
